Refactor Normals
